translation,bug fixes...
git-svn-id: https://pogeymon-online.svn.sourceforge.net/svnroot/pogeymon-online@547 a74332b3-73ba-4bb2-a0e2-b824b54a0f6f

+++ src/PokemonInfo/battlestructs.cpp
@@ -379,7 +379,7 @@ void TeamBattle::generateRandom()
         }
         while (p2.EVSum() < 510) {
             int stat = true_rand() % 6;
-            p2.setEV(stat, std::min(int(p2.EV(stat)) + (true_rand()%255), 255));
+            p2.setEV(stat, std::min(int(p2.EV(stat)) + (true_rand()%255), long(255)));
         }
 
         p.dvs().clear();
