+++ src/Utilities/functions.h
@@ -79,7 +79,7 @@ inline QString toBoldColor(const QString &mess, const QColor &col)
     return QString("<b><span style='color:%1'>%2</span></b>").arg(col.name(), mess);
 }
 
-inline int true_rand() {
+inline long int true_rand() {
     return rand() + (unsigned(clock())%RAND_MAX);
 }
